Split native projection infrastructure DTOs
