+++ src/app_core/actions/native_shell_dtos/automation.rs
@@ -1,0 +1,87 @@
+//! Automation snapshot DTOs for native-shell test and AI tooling.
+
+use radiant::gui::automation;
+use serde::{Deserialize, Serialize};
+use std::collections::BTreeMap;
+
+/// Stable semantic identifier for one automation node in the native shell tree.
+pub type AutomationNodeId = automation::AutomationNodeId;
+
+/// Quantized window-space bounds for one automation node.
+pub type AutomationBounds = automation::AutomationBounds;
+
+/// Semantic role describing how an automation node behaves in the GUI.
+#[derive(Clone, Copy, Debug, PartialEq, Eq, Serialize, Deserialize)]
+#[serde(rename_all = "snake_case")]
+pub enum AutomationRole {
+    /// Synthetic root of the automation snapshot tree.
+    Root,
+    /// Grouping container such as a panel or composite section.
+    Group,
+    /// Major panel surface.
+    Panel,
+    /// Toolbar or action strip.
+    Toolbar,
+    /// Tab-strip container.
+    TabList,
+    /// Toggleable tab node.
+    Tab,
+    /// Clickable button.
+    Button,
+    /// Search or text-entry field.
+    SearchField,
+    /// Slider or continuous meter interaction surface.
+    Slider,
+    /// Row in a list or table.
+    Row,
+    /// Table or row-hosting list surface.
+    Table,
+    /// Waveform interaction canvas.
+    WaveformRegion,
+    /// Map interaction canvas.
+    MapCanvas,
+    /// Focusable point inside the map canvas.
+    MapPoint,
+    /// Status/readout region.
+    Readout,
+    /// Dialog or modal container.
+    Dialog,
+}
+
+/// One node in the GUI automation tree.
+#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
+pub struct AutomationNodeSnapshot {
+    /// Stable semantic identifier for this node.
+    pub id: AutomationNodeId,
+    /// Behavioral role for this node.
+    pub role: AutomationRole,
+    /// Optional human-readable label shown by the GUI.
+    pub label: Option<String>,
+    /// Quantized window-space bounds.
+    pub bounds: AutomationBounds,
+    /// Optional current value or summary text.
+    pub value: Option<String>,
+    /// Whether the node is currently enabled.
+    pub enabled: bool,
+    /// Whether the node is currently selected or active.
+    pub selected: bool,
+    /// Stable action identifiers that this node can trigger.
+    pub available_actions: Vec<String>,
+    /// Additional deterministic metadata for AI/test consumers.
+    pub metadata: BTreeMap<String, String>,
+    /// Child nodes in semantic tree order.
+    pub children: Vec<AutomationNodeSnapshot>,
+}
+
+/// Full deterministic automation snapshot emitted for one GUI frame/state.
+#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
+pub struct GuiAutomationSnapshot {
+    /// Schema version for forward-compatible artifact readers.
+    pub schema_version: u32,
+    /// Quantized viewport width for the captured shell layout.
+    pub viewport_width: u32,
+    /// Quantized viewport height for the captured shell layout.
+    pub viewport_height: u32,
+    /// Root semantic automation node.
+    pub root: AutomationNodeSnapshot,
+}
